Adding dataVisualizer
