+++ src/dataVisualizer.js
@@ -1,0 +1,65 @@
+import * as d3 from 'd3';
+
+// sections :: [{name: String, count: Int, color: String}]
+export function donutChart(sections, size, divId) {
+  console.log('CHAMOU dataVisualizer');
+  const width = size;
+  const height = size;
+
+  const radius = 4 * size / 10;
+  const innerRadius = 2 * size / 10;
+
+  const numberTextSize = size / 10;
+  const nameTextSize = size / 40;
+
+  const arc = d3.arc()
+    .outerRadius(radius)
+    .innerRadius(innerRadius);
+
+  const pie = d3.pie()
+    .sort(null)
+    .value(d => d.count);
+
+  const svg = d3.select('div')
+    .attr('id', divId)
+    .append('svg')
+    .attr('width', width)
+    .attr('height', height)
+    .append('g')
+    .attr('transform', `translate(${width / 2}, ${height / 2})`);
+
+  const g = svg.selectAll('.arc')
+    .data(pie(sections))
+    .enter()
+    .append('g');
+
+  g.append('path')
+    .attr('d', arc)
+    .style('fill', d => d.data.color);
+
+  const gTexts = svg.selectAll('.arc2')
+    .data(pie(sections))
+    .enter()
+    .append('g')
+    .attr('class', 'arc');
+
+  gTexts.append('text')
+    .attr('transform', d => `translate(${arc.centroid(d)})`)
+    .style('text-anchor', 'middle')
+    .style('font-size', `${numberTextSize}px`)
+    .attr('font-family', 'consolas')
+    .text(d => d.data.count);
+
+  gTexts.append('text')
+    .attr('transform', (d) => {
+      const dAux = arc.centroid(d);
+      dAux[1] += (numberTextSize / 2);
+      return `translate(${dAux})`;
+    })
+    .style('text-anchor', 'middle')
+    .style('font-size', `${nameTextSize}px`)
+    .attr('font-family', 'consolas')
+    .text(d => d.data.name);
+}
+
+export default { donutChart };
